Add codec and 10-bit config in headset

+++ server/driver/configuration.h
@@ -50,7 +50,7 @@ struct configuration
 	};
 
 	std::array<encoder, 3> encoders; // left, right, alpha
-	std::optional<int> bit_depth;
+	std::optional<uint8_t> bit_depth;
 	std::optional<std::array<float, 3>> grip_surface;
 	std::vector<std::string> application;
 	bool debug_gui = false;
